Move retrieveTaskclusterToken to logic/taskcluster
